+++ src/main/java/roomescape/auth/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
