+++ src/main/java/roomescape/auth/LoginMemberArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
@@ -8,6 +8,7 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 import roomescape.AuthenticationException;
+import roomescape.member.MemberService;
 
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
